apps/voice-agent/pyproject.toml
[project]
name = "coziyoo-voice-agent"
version = "0.1.0"
description = "Coziyoo voice agent runtime"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.10.0",
  "fastapi>=0.115.0",
  "livekit-agents>=1.2.6",
  "livekit-plugins-noise-cancellation>=0.2.5",
  "livekit-plugins-silero>=1.2.6",
  "pydantic>=2.10.6",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.34.0"
]

[project.scripts]
coziyoo-voice-agent = "voice_agent.entrypoint:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
